% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kutz-package.R
\docType{package}
\name{kutz-package}
\alias{kutz}
\alias{kutz-package}
\title{kutz: Miscellaneous Materials including Datasets and Templates}
\description{
Package contains templates and data for use with R and Rmarkdown.
}
\author{
\strong{Maintainer}: Amos Okutse \email{okutseamos@gmail.com} (\href{https://orcid.org/0000-0001-6612-9788}{ORCID})

}
\keyword{internal}
